+++ src/rules/role-has-required-aria-props.js
@@ -28,7 +28,7 @@ module.exports = {
         }
 
         const roleValue = getElementAttributeValue(node, "role");
-        if (!roleValue) {
+        if (!roleValue || typeof roleValue !== "string") {
           return;
         }